backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.120.1",
    "mysqlclient>=2.2.7",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.27",
    "winuvloop>=0.2.0",
]
